Fixes to compile in Xcode 11 and run on iOS 14

+++ Classes/TouchHandlerView.m
@@ -65,9 +65,9 @@
         CGFloat relx = (locationInView.x - previousMouseLocation.x) / x_ratio;
         CGFloat rely = (locationInView.y - previousMouseLocation.y) / y_ratio;
 
-        if (fabsf(relx) < 1.0f)
+        if (fabs(relx) < 1.0f)
             relx = 0.f;
-        if (fabsf(rely) < 1.0f)
+        if (fabs(rely) < 1.0f)
             rely = 0.f;
 
         if (relx != 0.0f || rely != 0.0f)
